added interest simulation/main menu overhaul
The menu is now based off regular expressions and using if/else statements. With the regular expressions we are able to make more in-depth functions. The simulate will use a customer id, account id and the years to simulate for. The simulate function will first validate the existence of the customer, the account and also make sure the balance on the account is greater than 0. It will then perform the compound interest calculation and show the user the results.